requirements.txt
Flask==2.1.0
spotipy==2.19.0
python-dotenv==0.19.0